remove get_player_current_games_to_move to own module
