+++ chessdotcom/endpoints/player_current_games_to_move.py
@@ -1,0 +1,19 @@
+from ..client import Client, Resource
+from ..response_builder import ChessDotComResponse
+
+
+@Client.endpoint
+def get_player_current_games_to_move(
+    username: str, tts=0, **request_options
+) -> ChessDotComResponse:
+    """
+    :param username: username of the player.
+    :param tts: the time the client will wait before making the first request.
+    :returns: ``ChessDotComResponse`` object containing a list of Daily Chess games
+                where it is the player's turn to act.
+    """
+    return Resource(
+        uri=f"/player/{username}/games/to-move",
+        tts=tts,
+        request_options=request_options,
+    )
